refactor: resolve compilation errors and warnings
- Fix ref struct safety issues in move ordering
- Replace Move.Equals() with efficient Data property comparison
- Inline scoring logic to avoid ref safety violations
- Clean up unused parameters and duplicate initializations
- Update project to target .NET 8.0 with C# 14

+++ Meridian/Core/Evaluation.cs
@@ -104,7 +104,7 @@ public static class Evaluation
         score -= EvaluatePieces(ref board, Color.Black);
         
         // Add small random factor to avoid repetitions
-        score += (int)(board.FullMoveNumber * 7) % 10;
+        score += board.FullMoveNumber * 7 % 10;
         
         // Return score from the perspective of the side to move
         return board.SideToMove == Color.White ? score : -score;
